{{cookiecutter.project_name}}/requirements/base.txt
wheel==0.35.1
psycopg2-binary==2.8.6
alembic==1.4.3
SQLAlchemy==1.3.20
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4         # https://flask-sqlalchemy.palletsprojects.com/en/2.x/
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.1
flask-smorest==0.25.0           # https://flask-smorest.readthedocs.io/en/latest/quickstart.html
python-dotenv==0.15.0
flask-jwt-extended==3.25.0      # https://flask-jwt-extended.readthedocs.io/en/stable/
